+++ src/main/kotlin/bingsoochef/bingsoochef/global/error/GeneralExceptionHandler.kt
@@ -1,0 +1,32 @@
+package bingsoochef.bingsoochef.global.error
+
+import io.github.oshai.kotlinlogging.KotlinLogging
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+private val logger = KotlinLogging.logger {}
+
+@RestControllerAdvice
+class GeneralExceptionHandler {
+
+    @ExceptionHandler(IllegalArgumentException::class)
+    fun handleIllegalArgumentException(e: IllegalArgumentException): ResponseEntity<ErrorResponse> {
+        return ResponseEntity<ErrorResponse>(
+            ErrorResponse(HttpStatus.BAD_REQUEST.value(), e.message),HttpStatus.BAD_REQUEST)
+    }
+
+    @ExceptionHandler(CustomException::class)
+    fun handleCustomException(e: CustomException): ResponseEntity<ErrorResponse> {
+        return ResponseEntity<ErrorResponse>(
+            ErrorResponse(e.getStatusCode(), e.message),e.httpStatus)
+    }
+
+    @ExceptionHandler(NotFoundException::class)
+    fun handleNotFoundException(e: NotFoundException): ResponseEntity<ErrorResponse> {
+        return ResponseEntity<ErrorResponse>(
+            ErrorResponse(e.getStatusCode(), e.message),e.httpStatus)
+    }
+
+}
